requirements.txt
streamlit>=1.35.0
google-cloud-bigquery>=3.11.0
pandas>=2.0.0
plotly>=5.18.0
db-dtypes>=1.1.1
pyarrow>=12.0.0
anthropic>=0.89.0
